C++: Fix build against Yocto SDKs
When using `. environment-setup-*` against a Yocto SDK, we end up with PKG_CONFIG_PATH set to the target sysroot and the pkg-config binary coming out of Yocto SDK and always locating packages in the target.

That means the build of slint-compiler fails as it needs font-config from the host system, when enabling support for embedding glyphs.

If you need the slint-compiler for a C++ app in a Yocto environment *with* support for glyph embedding, then the compiler needs to be built separately first as a host build and then injected. That's what we do in your Yocto recipes anyway. To keep an innocent build against a Yocto SDK working though, disable the software-renderer feature in that scenario.